+++ src/classes/extensions/index.mjs
@@ -1,0 +1,53 @@
+/**
+ * @namespace extensions
+ */
+
+/**
+ * @namespace add-ons
+ */
+
+import ObjectExt from "./ObjectExt.mjs"
+import StringExt from "./StringExt.mjs"
+import NumberExt from "./NumberExt.mjs"
+import DateExt from "./DateExt.mjs"
+import FunctionExt from "./FunctionExt.mjs"
+import ArrayExt from "./ArrayExt.mjs"
+import ArrayBufferExt from "./ArrayBufferExt.mjs"
+import SharedArrayBufferExt from "./SharedArrayBufferExt.mjs"
+import TypedArrayExt from "./TypedArrayExt.mjs"
+import SetExt from "./SetExt.mjs"
+import ScreenExt from "./ScreenExt.mjs"
+import LocationExt from "./LocationExt.mjs"
+import HTMLElementExt from "./HTMLElementExt.mjs"
+import HTMLImageElementExt from "./HTMLImageElementExt.mjs"
+import ImageExt from "./ImageExt.mjs"
+import DOMPointExt from "./DOMPointExt.mjs"
+import DOMRectExt from "./DOMRectExt.mjs"
+import DOMMatrixExt from "./DOMMatrixExt.mjs"
+import CSSStyleSheetExt from "./CSSStyleSheetExt.mjs"
+import ShadowRootExt from "./ShadowRootExt.mjs"
+import HTMLDocumentExt from "./HTMLDocumentExt.mjs"
+
+export {
+	ObjectExt,
+	StringExt,
+	NumberExt,
+	DateExt,
+	FunctionExt,
+	ArrayExt,
+	ArrayBufferExt,
+	SharedArrayBufferExt,
+	TypedArrayExt,
+	SetExt,
+	ScreenExt,
+	LocationExt,
+	HTMLElementExt,
+	HTMLImageElementExt,
+	ImageExt,
+	DOMPointExt,
+	DOMRectExt,
+	DOMMatrixExt,
+	CSSStyleSheetExt,
+	ShadowRootExt,
+	HTMLDocumentExt
+}